requirements.txt

# Core AI and Web Framework
langchain
langchain-huggingface
streamlit
langchain_community
# Vector Store & Embeddings
faiss-cpu
sentence-transformers
accelerate

# Utilities
python-dotenv
requests

huggingface_hub

# Testing
pytest
pytest-mock
requests-mock